+++ lib/executor/state.rs
@@ -99,6 +99,11 @@ impl State {
                 self.symbolize_expression(lhs)?,
                 self.symbolize_expression(rhs)?,
             )?,
+            #[cfg(feature = "il-expression-ashr")]
+            il::Expression::AShr(ref lhs, ref rhs) => il::Expression::ashr(
+                self.symbolize_expression(lhs)?,
+                self.symbolize_expression(rhs)?,
+            )?,
             il::Expression::Cmpeq(ref lhs, ref rhs) => il::Expression::cmpeq(
                 self.symbolize_expression(lhs)?,
                 self.symbolize_expression(rhs)?,